Update Navigation component to spec
The Navigation component is now included on every page automatically.
It's styled according to the Figma spec here:
https://www.figma.com/file/oXxyGrHDtWrjvx1oeiHZKN/Untitled?node-id=181%3A194

We preload the variable truetype font definition, but also fall back to
the standard non-variable woff2 format if the browser doesn't support
variable fonts.

Also updates the Burger component to use an actual button, and to match
the spec visually.