Fix up special handling for index page
Some of the previous special handling assumed that
it was legal to have an empty slug for a page, but
since slug is the identifier field for a page, it
can't be empty. Instead, we just special case the
page with slug === 'index' to support proper
routing to and from '/'.