Move admin page into next.js pages/
The admin page was defined as a single static HTML page. This is pretty
limiting in what we can do with it; it means that we don't get any
benefits from Typescript or our build system when configuring our CMS.

As our CMS config gets more complex, we'll likely want to import React
components and other code, which we couldn't do with the old setup. This
strategy is much more flexible.