(Partly) resolve flash of incorrect styles
next-mdx-remote wraps its content in a div on the server side, which
means that our styles targeting children of `main` broke on first page
load/when JS was disabled. This commit adds a wrapper around calls to
`hydrate` that append a class that can be targetted for styling, and
updates our global styles to handle that class when it exists.